Fehlerbehebung bei Verhältnisberechnung.
git-svn-id: https://invertika.googlecode.com/svn/trunk@4844 784bf6b6-0765-11df-962c-b5c252e50527

+++ Invertika Editor/FormCreateMapThumbnailsAndMinimaps.cs
@@ -241,7 +241,7 @@ namespace Invertika_Editor
 
 				int imageSizeOriginalWidth=(int)pic.Width;
 				int imageSizeOriginalHeight=(int)pic.Height;
-				double imageVerhaeltnis=imageSizeOriginalWidth/imageSizeOriginalHeight;
+				double imageVerhaeltnis=imageSizeOriginalWidth/(double)imageSizeOriginalHeight;
 
 				//int imageSize=6400;
 				int imageSize=800;
